refactor: 인터셉터를 통한 HTTP Error 핸들링
- 패스워드가 맞지 않는 경우, http 에러를 서비스 로직에서 날리는 것이 맞는가?
- single responsibility principle을 위반했다고 판단하고, service 로직에서 IncorrectPasswordException(커스텀) 에러를 던지면, controller에 붙어있는 interceptor 데코레이터에서 인식 후 Forbidden으로로 변환해 에러를 표시한다.
- 이렇게 함으로써, 책임과 역할을 더 잘 분리해서 에러를 핸들링할 수 있다.

+++ src/posts/posts.service.ts
@@ -10,7 +10,10 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { PostEntity } from './entities/post.entity';
 import { Repository } from 'typeorm';
 import { BCRYPT_SALT } from '../environments';
-import { EntityNotFoundException } from '../common/exceptions/exception';
+import {
+  EntityNotFoundException,
+  IncorrectPasswordException,
+} from '../common/exceptions/exception';
 
 @Injectable()
 export class PostsService {
@@ -62,8 +65,10 @@ export class PostsService {
   }
 
   /**
-   * @param {number} id 게시물의 id
-   * @param {string} enteredPassword 게시물 삭제 권한 확인을 위한 패스워드
+   * @param {number} id
+   * @param {string} enteredPassword
+   * @throws {EntityNotFoundException}
+   * @throws {IncorrectPasswordException}
    * @returns {Promise<boolean | undefined>}
    */
   async delete(
@@ -76,24 +81,34 @@ export class PostsService {
       throw new EntityNotFoundException();
     }
 
-    await Promise.all([this.validatePassword(enteredPassword, post.password)]);
+    if (!(await this.isCorrectPassword(enteredPassword, post.password))) {
+      throw new IncorrectPasswordException();
+    }
 
     return (await this.postsRepository.delete({ id: id })) && true;
   }
 
+  /**
+   *
+   * @param {CreatePostDto} post
+   * @returns {Promise<void>}
+   * @private
+   */
   private async transformPassword(post: CreatePostDto): Promise<void> {
     post.password = await bcrypt.hash(post.password, BCRYPT_SALT);
     return Promise.resolve();
   }
 
-  private async validatePassword(
+  /**
+   * @param {string} enteredPassword
+   * @param {string} postsPassword
+   * @returns {Promise<boolean>}
+   * @private
+   */
+  private async isCorrectPassword(
     enteredPassword: string,
     postsPassword: string,
-  ) {
-    if (!(await bcrypt.compare(enteredPassword, postsPassword))) {
-      throw new ForbiddenException(
-        "you don't have permission to access this resource",
-      );
-    }
+  ): Promise<boolean> {
+    return await bcrypt.compare(enteredPassword, postsPassword);
   }
 }
